# Support fast goroutine ID access on arm64