--- a/goid_go1.5_arm.s
+++ b/goid_go1.5_arm.s
@@ -14,7 +14,6 @@
 // for names of contributors.
 
 // Assembly to mimic runtime.getg.
-// This should work on arm64 as well, but it hasn't been tested.
 
 //go:build arm && gc && go1.5
 // +build arm