--- a/goid_go1.5_arm.s
+++ b/goid_go1.5_arm.s
@@ -16,8 +16,10 @@
 // Assembly to mimic runtime.getg.
 // This should work on arm64 as well, but it hasn't been tested.
 
+//go:build arm && gc && go1.5
 // +build arm
-// +build gc,go1.5
+// +build gc
+// +build go1.5
 
 #include "textflag.h"