--- a/runtime_gccgo_go1.8.go
+++ b/runtime_gccgo_go1.8.go
@@ -1,3 +1,4 @@
+//go:build gccgo && go1.8
 // +build gccgo,go1.8
 
 package goid