--- a/runtime_go1.5.go
+++ b/runtime_go1.5.go
@@ -13,6 +13,7 @@
 // permissions and limitations under the License. See the AUTHORS file
 // for names of contributors.
 
+//go:build go1.5 && !go1.6
 // +build go1.5,!go1.6
 
 package goid