--- a/runtime_go1.9.go
+++ b/runtime_go1.9.go
@@ -1,3 +1,4 @@
+//go:build gc && go1.9
 // +build gc,go1.9
 
 package goid