This commit updates the library to support fast access to the current goroutine ID on arm64 architectures. I noticed the need for this when running a few benchmarks in cockroachdb's `tracing` package.

Unfortunately, the existing 32-bit ARM assembly does not work with arm64, because it uses a MOV instruction variant with an incorrect width (32-bit vs. 64-bit). Replacing MOVW with MOVD does the trick. See https://pkg.go.dev/cmd/internal/obj/arm64#hdr-Instructions_mnemonics_mapping_rules.

I don't think CI tests on arm64, so here's some confirmation that this all works on my M1 mac:
```
➜ go env
GO111MODULE=""
GOARCH="arm64"
GOBIN=""
GOCACHE="/Users/nathan/Library/Caches/go-build"
GOENV="/Users/nathan/Library/Application Support/go/env"
GOEXE=""
GOEXPERIMENT=""
GOFLAGS=""
GOHOSTARCH="arm64"
GOHOSTOS="darwin"
GOINSECURE=""
GOMODCACHE="/Users/nathan/Go/pkg/mod"
GONOPROXY=""
GONOSUMDB=""
GOOS="darwin"
GOPATH="/Users/nathan/Go"
GOPRIVATE=""
GOPROXY="https://proxy.golang.org,direct"
GOROOT="/Users/nathan/Go/go"
GOSUMDB="sum.golang.org"
GOTMPDIR=""
GOTOOLDIR="/Users/nathan/Go/go/pkg/tool/darwin_arm64"
GOVCS=""
GOVERSION="go1.17.3"
GCCGO="gccgo"
AR="ar"
CC="clang"
CXX="clang++"
CGO_ENABLED="1"
GOMOD="/dev/null"
CGO_CFLAGS="-g -O2"
CGO_CPPFLAGS=""
CGO_CXXFLAGS="-g -O2"
CGO_FFLAGS="-g -O2"
CGO_LDFLAGS="-g -O2"
PKG_CONFIG="pkg-config"
GOGCCFLAGS="-fPIC -arch arm64 -pthread -fno-caret-diagnostics -Qunused-arguments -fmessage-length=0 -fdebug-prefix-map=/var/folders/hn/4tk2rm215ld9k2x8yq7rr78m0000gq/T/go-build3083548140=/tmp/go-build -gno-record-gcc-switches -fno-common"

➜ GO111MODULE=off go test . -count=1 -v
=== RUN   TestGet
--- PASS: TestGet (0.00s)
PASS
ok  	github.com/petermattis/goid	0.099s

➜ GO111MODULE=off go test . -bench=. -benchmem
goos: darwin
goarch: arm64
pkg: github.com/petermattis/goid
BenchmarkGet-10    	550983054	         2.268 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/petermattis/goid	1.583s
```

Here's the impact this change has on performance:
```
➜ GO111MODULE=off benchdiff .
checking out '7abefc4'
building benchmark binaries for '7abefc4' 1/1 -
checking out 'c091d1c'
building benchmark binaries for 'c091d1c' 1/1 |

  pkg=1/1 iter=10/10 petermattis/goid \

name    old time/op    new time/op    delta
Get-10    2.21µs ± 1%    0.00µs ± 4%   -99.91%  (p=0.000 n=9+9)

name    old alloc/op   new alloc/op   delta
Get-10     65.2B ± 2%      0.0B       -100.00%  (p=0.000 n=10+10)

name    old allocs/op  new allocs/op  delta
Get-10      1.60 ±38%      0.00       -100.00%  (p=0.000 n=10+10)
```