@tamird 2021-12-23
Looks good, but I think CI is entirely busted; RIP travis.

Can I interest you in some yak shaving? https://github.com/marketplace/actions/setup-go-environment