@nvanbenschoten 2021-12-23
FWIW, the other yak to shave with this project is go module support. The project has no external dependencies, so it doesn't need any vendoring, but the `GO111MODULE=off` stuff is increasingly (since 1.16) frustrating to deal with. I don't know how much we care about pre-Go 1.11 support anymore.