@tamird 2021-12-23
Added GH Actions CI in #13. Mind rebasing? Then we'll get CI here (though we don't test against ARM).