@nvanbenschoten 2021-12-23
Rebased. Thanks for setting that up. Too bad GH Actions doesn't come with arm support out of the box.